requirements.txt
fastapi
pydantic
supabase
langchain_google_genai
langchain
python-dotenv